# src/main/resources/api-specs/openapi.yml
openapi: 3.0.1
info:
  title: WEB Java course API documentation
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:7070
  description: Local server
paths:
...
  /v1/api/orders/{id}:
    get:
      tags:
      - order-controller
      operationId: getOne_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDto"
    put:
      tags:
      - order-controller
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderForm"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDto"
    delete:
      tags:
      - order-controller
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
...
  /v1/api/orders:
    get:
      tags:
      - order-controller
      operationId: getAll_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderDto"
    post:
      tags:
      - order-controller
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderForm"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDto"
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    ProductDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
        categories:
          uniqueItems: true
          type: array
          items:
            $ref: "#/components/schemas/Category"
    OrderForm:
      required:
      - productId
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        quantity:
          maximum: 100
          minimum: 1
          type: integer
          format: int32
    OrderDto:
      type: object
      properties:
        product:
          $ref: "#/components/schemas/ProductDto"
        quantity:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
...